+++ app/models/step.rb
@@ -24,24 +24,7 @@ class Step < ApplicationRecord
     super
   end
 
-  def options
-    raw.dig("fields", "extendedOptions") || legacy_options(super)
-  end
-
   def options_list
     options.map { |hash| hash["value"] }
-  end
-
-  # TODO: Remove this migration step when extendedOptions is ready to be solely
-  # relied upon in all environments.
-  # Convert the legacy 'options' field into the same structure as extendedOptions
-  # so the view only needs to support one data structure:
-  #
-  # [{"value"=>"Foo"}, {"value"=>"Bar"}] rather than ["Foo", "Bar"]
-  private def legacy_options(legacy_options)
-    return nil if legacy_options.nil?
-    legacy_options.inject([]) do |array, legacy_option|
-      array << Hash["value" => legacy_option]
-    end
   end
 end